requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
google-auth
google-auth-httplib2
google-api-python-client
python-dotenv
fitz
PyMuPDF
openai
